dpmi: check psp on int23 [#1438]
If int23 requests the termination, we need to decide whether
or not to kill the entire DPMI client. We do not want comcom32
to be always killed by ^C together with some app that it runs.
We only want that app to be killed and comcom32 to keep working.
For that I compare the current and initial DPMI psp. If they
do not match, then we assume that some other program was started,
and so we do not terminate the entire DPMI app which would be
comcom32. We only terminate the newly started RM program then.