coopth: introduce coopth_create_vm86() fn
This function allows to start the thread directly from realmode
code, removing the need of allocating the hlt trampoline.
Well, in fact it is just a wrapper that allocates the trampoline
internally. But it saves the user from writing the quite non-intuitive
starter code.

fdpp/hooks.c and msdoshlp.c are adapted to this new API (another patch -
another API).

*_custom() API is removed as it required the too verbose code from
user. But internally all the code is kept, we just have a convenient
wrapper now.