msdoshlp: fix race with rmreg buffer [#1551]
It appears that there is a small race at msdos coopthread
start-up when the buffer can be re-used by HW interrupt
before the first instruction is executed, which is usually
int 31h. As soon as the first instruction is executed, the
buffer can be freely re-used, which is why initially I thought
its safe.

Instead of using separate buffers for each thread, this patch
just goes for more careful usage. Namely, the local buffer is
used for as long as possible, and the data is copied to the
global buffer only when it is actually going to be used. No
more "false" uses of the global buffer.
Also added the refcount for global buffer, and the error will
be printed if it is ref'ed more than once.