signal: inject timer before keypresses [#1580]
Timer has the highest prio, so if it is injected, no other IRQ
will be injected. But if we inject some other IRQ first, then
timer can still be injected, triggering the race in some programs.

This patch makes SDL to not inject kbd irq before timer irq
(sdl injects irq from update_screen() call).

This patch is "not needed" because such condition is already
disallowed in PIC by prev hack. But this change makes sdl plugin
to inject irqs more similar to an X plugin that does it after timer.