dpmi: make sti/iret atomic
sti/iret is the only way to restore IF flag because IF is ignored
by iret. sti holds interrupts for the next instruction. We need to
do that too.

This is a major fix, and the need of masking timer IRQs by DPMI
may not be unneeded in some cases.