mfs: make IO ops atomic wrt region locking
Region locking in linux is advisory so we check the locks by hands.
This is race-prone as someone may change the locking after we checked
it but before we performed an IO.
Use extra exclusive locking to protect against such conditions.

Note: the obvious "fix" here would be to just protect the IO ops
with the region lock that matches the requested IO region. That
would remove the possibility of a partial IO (if the beginning of
the IO region is not locked), but DOS doesn't allow partial IO either,
so that would seem to be still fine.
Unfortunately such strategy won't work on RO-opened files, as the
exclusive region locks require RW mode...