vtmr: add smi handler [fixes #1607]
That handler (just a coopthread with async context) checks if
vtmr is masked, and if so - requests the timer interrupts directly.

This fixes WP62 that masks all irqs except timer.
Fortunately SMI ignores any masking.

So as a matter of fact, this vtmr subsystem is not free of troubles.
It required lots of bindings to DPMI, and now also smi...