dpmi: DPMI_interrupt should distinguish jumps [#1610]
DPMI_interrupt was assuming it is always called to.
But if it is instead jumped to, free PM stack and immediately go to RM,
bypassing do_dpmi_int() machinery.
It appears most progs actually jump to prev handler, instead of doing
call. win31, Transport Tycoon, CMUD and other progs I tested, hit that
code path now. So it can change a lot and should be properly tested.