CI: Disable share tests that provoke failures
1/ MFS doesn't follow the same semantics for delete / setfattrs when
referencing a file opened with SH_COMPAT. MS-DOS 6.22 share has
different behaviour whether the open file was opened by the same process
requesting the delete/setfattrs or not. MFS now allows deletion in both
cases regardless.

2/ FDPP's internal share operating on FAT does not respect the above
either, but it needs to be compatible with MS-DOS 6.22's share. For now
disable the failing tests only on FDPP, they can be reinstated as the
problem is fixed.