conf: add $_fluid_volume option [#1604]
Volume can normally be controlled by the SB mixer.
But we need this option too, as the lower values here avoid
distortion. So we need to set $_fluid_volume to some low value
and increase the volume with SB mixer, but not everyone has
SB mixer installed. So we need 2 knobs.