add -kernelcommandline switch for RxDOS.3, FreeDOS
Upcoming patches to ldosboot and the FreeDOS kernel allow
to receive a kernel command line when loaded under the
FreeDOS load protocol. (RxDOS.3 load protocol has had
support for the kernel command line for a while already.)

The script src/dosemu seems to run into problems when the
-kernelcommandline switch's contents contain double quotes.
This is shared by the -input switch's handling, which was
largely used as the base for this commit.

The behaviour of the kernel command line depends on the
particular kernel:

* RxDOS (with lDOS init) does not yet use it.

* testpl displays the contents.

* lDebug enters the contents into its command line buffer,
with unescaped semicolons converted to linebreaks.

* The FreeDOS kernel patch being prepared will allow to
specify up to three different CONFIG.SYS filenames that
are tried in order. (This is useful to quickly switch a
particular dosemu2 instance to another configuration file
without having to modify any files or impacting other
uses of the same drive C: on the system.)

The string "@none@" is handled specifically to indicate
that no command line should be passed. (This differs from
passing an empty command line, as the kernel can tell. In
particular, lDebug will load its startup script file if it
receives no command line.)

The protocol for passing the command line is based on how
existing bootable lDebug handles a BOOT command with the
PROTOCOL=FREEDOS and a CMDLINE=1 switch, or with either of
the RxDOS.3 or lDOS protocols which allow a command line by
default. The FreeDOS extension therefore mirrors the stack
frame layout used for the lDOS load protocol.