dpmi: remove dpmi_return_request()
It was used to return from DPMI by a signal.
That was mainly needed to call handle_signals() and pic_run().
But with new pic we don't need to return for pic_run(), and
handle_signals() can be called from dpmi directly.

Hope this speeds things up.
Quite an experimental change.