msdos: avoid hooking its twice [fixes #1677]
When ints are inherited from prev client, msdos is already in
an int chain. Hooking ints again and chaining to prev handler,
causes it to take controll again, on behalf of prev client.
But since the client number is not known at that point, it enters
the infinite loop.

The bug was triggering only when someone else hooked int21, because
otherwise the msdos ints were not inherited... That was needed
for vtmr. So changed the technique the way only the vtmr interrupts
are not inherited.