pic: read poll only with addr==0
vtmr thread reads IMR and can reset poll mode as a side effect.
Either we need this hack or implement some critical sections.
Critical section is quite difficult with vtmr (currently coopth
is disabled there) and is difficult for vint (no coopth there
currently and unclear what to do with dpmi binding).

So lets instead "extend" the PIC in a way that would keep things
manageable.