dpmi: remove djgpp's dpmi.h
Just rip off the parts that we need.
But we should avoid providing djgpp-compatible prototypes, which
is what this patch removes. It was intended for something that isn't
happening. Namely dpmistub doesn't seem to happen. If we ever bind
djgpp here then it should talk to coopth directly (as does fdpp)
rather than via dpmi stub.