Revert "dpmi: do not inherit IDT when RSP is installed [#1341]"
This reverts commit 5e35fae6d5839d18b6303e0ec54ade75b2f0aaa2.

This breaks msdos in presence of RSP because it doesn't know
the vectors were not inherited. So it doesn't install handlers
and nothing works.

Instead we need to help RSP and msdos (with some API extensions)
finding out whether the IDT was inherited or not. This should be
in a subsequent patch.