xms: clean up himem hooking
Currently if our built-in xms only works as umb driver, it skips
the external himem callout by incrementing esp.
This is quite a nasty thing to do so at least lets move it into
an asm driver code.
Adjust hook stub to not call directly into dosemu, but rather
to call our trampoline that can do the checking. Flag is passed
in ZF.