xms: implement top-down alloc strat [closes #1851]
This avoids possible collisions with HX's fixed allocs to 4Mb space.
XMS maps should be confined within low 16Mb to stay in reach for DMA.
But this doesn't mean XMS is now limited to 16Mb or alike, its just
a mapping window. XMS can be of any size, as before.