fp87-sim: reorganize masked exception handling
We now keep the masked exception state on the native FPU except if
access to the status word is needed (FSTSW/FSTENV/FSAVE)
and if some emulator code can cause exceptions, in which case
we need to save before and clear after.

This fixes all remaining mcpdiag failures (except one, but that also fails
on real hardware), and QEMU's test-i386-fp-exceptions.c test.