simx86: avoid nodes overlap again [#1876]
Checking querymark at PC appears not enough because PC can increment
by multiple bytes. So if we see the mark, it doesn't mean we are on its
beginning. If we don't see the mark, it doesn't mean we haven't missed
one already.
So check entire range before closing a new node.

Add assert to e_markpage() to never over-mark the mark.