KVM: remove use of MAPPING_KVM_UC
This was an attempt at fixing console graphics under KVM but that doesn't work
and is inactive here anyway. I checked console graphics and it works for me
under vm86 and cpuemu, but indeed not KVM, but think this is related to port I/O.
If we need to bypass caching we can always resurrect this and make sure it's
active (by calling mprotect_kvm from kmem_map_single).