KVM: map monitor directly after main pool map.
The use of DOSADDR_REL for the monitor was a little awkward since it assumes that
monitor is allocated within a 32bit distance of mem_base. In practice it would
wrap around (allocated below mem_base) and pick some random high address in 4G RAM.

It's cleaner to have a defined location, e.g. directly behind the main pool,
so just do that.