mapping: allow $_dpmi=(0).
This is related to VCPI (#1880) but this should work no matter what.
Allocate enough memory to allow for EXTMEM, XMS, and VGAEMU LFB mappings
independent of DPMI settings: if $_dpmi=(0), we reserve config.dpmi_base
memory (default: 32MB).

Relatedly this cleans up config.vgaemu_memsize a bit so it doesn't need to
be checked for 0 (which means 1MB for historical reasons) every time.