serial: avoid shadowing problems [#1908]
Com port pointer was always passed under name "com", but there
is also a global array with the same name. That was producing the
difficult to debug problems. Eg when some macro uses com[i] hoping
to access a global array, but instead touching the local var.

This patch tries to globally change "com" to "c". But unfortunately
the missed conversions would not give a warning. :(
So this can regress.