fdpp: run userhook.sys from -K drive [#1913]
This patch adds support for fdpp's new shared part of boot
parameters block. With that shared part it is possible to pass
some info to fdpp during boot.
Now when -K drive is created, it is passed to fdpp in a hope
fdpp is still parsing fdppconf.sys.

As a result, userhook.sys is now looked up on -K drive.
C:\userhook.sys is not parsed if -K is used.
Add another hook called globhook.sys. It is called from C: even
if -K is used.