xms: avoid external himem hang with $_dos_up=(2)
With $_dos_up=(2) fdpp is already in HMA when the XMS driver is
loaded. External himem would trash fdpp in HMA and hang the system.
So when himem does the installation check, report that the XMS
driver is already installed. This prevents himem from loading.
But use the special code that our driver can recognize as "not
installed".