mfs: in FindFirst/Next return large file's size as 0xFFFFffff
Prior to this the size would be returned as modulo 4 GiB.
Fix to use the set_32bit_size_or_position function which
will clamp the size to the maximum value instead.

Test: Create a file like so:
dd if=/dev/urandom bs=1024 seek=4194304 count=2 of=big1.dat

(This creates a sparse file that is just over 4 GiB.)

Prior to this commit FreeCOM shows the file size as
2048 bytes. With this commit this improves to show
the size as 4,294,967,295 bytes.