mapping: Use pgalloc for file-based SHM, pass target (#1934)
Using the page allocator instead of smalloc allows us to mmap memory
on demand, but still keeping track of the file offsets that were used
to map that memory.

The realloc method was renamed to resize, as we can keep both the
mmap area and file area the same without copying then, but if resizing
isn't possible, mapping.c code deals with it, using a new allocation
and copying to it.

Passing target will makes it easier to align to huge pages.