conf: add $_local_dir and $_dosemurc opts [#1914]
... and set them to "~/.dosemu/dosemurc" by default.
The problem is that ~/.dosemu/.dosemurc violates FHS; it suggests to
not use leading dot except when the config file is directly at $HOME.
To avoid further problems, this patch just makes it configurable.
The old locations are also searched, but if found - they are used
after printing an error.
You are free to ignore that error, or you can update the name, or
you can change the setting in dosemu.conf.