smalloc: align size properly to find free space
There was an issue where in a pool of 0x02000000, after allocating 0x110000
at the bottom, allocating the remaining 0x01ef0000 from the top via
smalloc_aligned_topdown failed because it looked for a block of size 0x01ef0fff.
Instead of simply adding align-1 to size we should align the same way PAGE_ALIGN,
and also feed that value that the OOM error print handler to avoid a prio==-1
assert.