signal: fix i386 true_vm86() / JIT FPE fault handling.
0ba74dec introduced minsigsegv() but that only dealt with JIT SIGSEGV. It
also need to trap SIGFPE for divisions by zero.

Moreover it needs to also deal with real vm86 for SIGFPE, SIGSEGV, SIGILL,
SIGTRAP and SIGBUS (the last two maybe not depending on kernel behaviour,
but it doesn't hurt to check).

Since minsigsegv is no longer about SIGSEGV only, renamed it to minfault.