mapping: allow register_hardware_ram_virtual with vbase=-1
Up so far we've always had a vbase for virtual hardware, but with
extmem/XMS the vbase is only needed for DPMI.

Allowing vbase=-1 makes it possible to use a minimal main_pool of
just 0x110000 if $_dpmi=(0), since XMS/EMS don't need to be mapped
from main_pool then.

This changes do_get_hardware_ram to always return the pointer to
the struct and set the virtual address optionally, where it was
the opposite before, as we now sometimes only need that pointer but
not the virtual address.