mapping: reintroduce static aliasmap for lowmem only
As mentioned in commit 0c983365, commit 705dab48 made dosaddr_to_unixaddr
much slower. While sim is fast again, JIT was still slow in cpatch code.
This patches reintroduces a static lowmem_aliasmap, shared by all
lowmem hardware RAM registrations; for high memory it goes via the
malloc'ed per-mapping aliasmaps.

The test in 0c983365 (which spends almost all its time in the cpatch)
went from 6.5 to 4.5 seconds for JIT. It's harder to make the cpatch
use the page cache since it needs to check for code at the subpage level,
and the test writes to data in the same page as the code.