vgaemu: use cpusim instead of instremu
Introduce instr_emu_sim(), which behaves like instr_emu() but
uses simx86' simulator mode.

So it shortcuts VGA port i/o, has a countdown to exit, and if
an instruction causes an exception it simply exits and lets the
caller redo it using the regular backend.

As this can emulate more instructions than the old instr_emu, the number
of faults is reduced, particularly in 32bit code. E.g. for DOOM this
number was reduced by about 66%.