cpuemu: don't also check TheCPU.sigalrm_pending in CeS_INSTREMU check
Fixes performance regression from #1975 / e056787b

It checked for it but didn't reset which caused performance issues (DOOM
ran much more slowly). It's better to just let the existing sim/interp logic
check for it and exit the interpreter loop that way when an async signal
arrives.

Also set TheCPU.sigalrm_pending = 0 when entering cpuemu to make sure it's
not stuck there from a previous call and prevent the above situation
completetely.