KVM/vgaemu: use MMIO region for planar i/o.
Note that when exiting for MMIO you must return to KVM immediately for a
clean exit. The clean exit can either be done using a signal that is only
unblocked during KVM_RUN, or more simply, on newer kernels using
run->immediate_exit=1. This patch only implements the newer method.