dpmi: extend 0x503 to deal with shms
We used to allow applying normal free (0x503) to shmem blocks
because Syndicate game does this. Lets make an extension out of
that, which can be used to unlink shm.

msdos_ldt: use that (and previous UnmapHWRam) extensions to unlink
shm early on start-up. This avoids occasional shm leaks in case of
dosemu process crash.