mfs: replace dir lock with shlock API [#1345]
This patch introduces the shlock library that implements the
shared and exclusive locking on top of UUCP-style locks+flock().

The very "clever" (and convoluted) usage of that API allowed
to get rid of a directory-based range locks. The shlock lib itself
uses only flock() locks, also on directories, but the range locks
on a directory are now completely avoided.

The share mode emulation port to shlock API is in todo.