msdos_ldt: generate unique shm name [#2020]
If dpmi impl is changed to 1.0 then shms will no longer be unique
per-dosemu2 process, as dpmi-1.0 specifies the communication between
VMs via shm. So all shms that needs to be unique, should be made so
explicitly.

utilities: add tempname() function to create temp names, similar
to mktemp(). mktemp() is a deprecated function because it checks
for the file existence when generating the unique name. We do not
check anything, using O_EXCL instead.

dpmi: add SHM_EXCL flag to pass O_EXCL to shm_open().