reintroduce timerfd-based timers again [#506]
timer_create() is a rather broken API, and SIGEV_THREAD is
broken up to not being usable at all:
https://sourceware.org/bugzilla/show_bug.cgi?id=30558

Use timerfd by default, leaving the broken impl for freebsd only.
This patch re-introduces the timerfd code in a rewritten form:
it no longer relies on an ioselect machinery. Therefore moved
to libs.