midi: get synth type from producer dev [#2026]
Now the device can request the particular synth type.
If it is not available, the default one is used.
Unfortunately the "unavailable" situation is rare as the alsa
plugin registers the interfaces for both GM and MT32.

This patch changes sb-midi to always request GM.
I really doubt anyone used MT32 with sb-midi.
In the future I also plan to bind sb's mpu401 to GM as well.