build procedure: use .incbin instead of objcopy bin->elf (#721)
As explained in https://sourceware.org/bugzilla/show_bug.cgi?id=30561
one can use .incbin to include a binary blob in assembly instead of
using $(OBJCOPY) (which can't convert to Mach-O or Windows PE).

Adding this as an asm section to $@_offsets.c additionally
eliminates the need for $(HOST_LD) -r.

Next the _binary_*_o_bin_size fields are not used so were removed,
and additionally an asm name is added for macOS (since otherwise it's
prefixed with a _ there).