mapping: strip PROT_EXEC when possible
macOS doesn't like PROT_EXEC on shared memory so to fix this properly,
remove it anywhere where possible. We only need it for native
and vm86. JIT sets PROT_EXEC on the codegen buffer via dlmalloc; KVM
only needs it in the page tables inside the VM, not on the backend
memory.

When double checking with native DPMI I found that ROM was marked
noexec, so corrected that.