cpu: check if FPU IRQ can be injected [#1886]
Some program may rely on IGNNE being set by the port 0xf0, and
disable interrupts before raising an FPU exception. In that case
our fnclex work-around in bios.S is not going to be called so we
can get stuck in a SIGFPE flood.
Just terminate dosemu is that case.
Other options may include the force unmasking, or just call the
inthandler directly, or clear the pending exception in an FPU state.
Since the problem is fairly theoretical, use simplest approach, as
was discussed with Bart.