vtmr: introduce latch handler [fixes #2077]
That handler inspects ISR to see if the latching happens in an
IRQ context or not. If not and an underflow is seen during latching,
then we advance the timer tick.

This replaces the isset_IF_async() hack that is not very reliable.
After all, IF is always clear in an IRQ handler.