mouse: distribute virq calls [fixes #2076]
... across timer interrupts.

Can be enabled with emumouse t 1

It seems gunmetal game can't queue multiple mouse events.
It processes them in a timer interrupt, so we need to distribute
the accumulated mouse events across the timer interrupts.

The alternative, more proper fix, may be to throttle the mouse
events with the use of evtimer. But too heavy throttling breaks
the game's ability to evaluate the mouse move accelleration, while
too light throttling does not always allow the timer interrupt to
sneak in.