shlock: simpler implementation
Instead of a single lockfile, use directory with multiple lockfiles.
Directories has an advantage over files, that they are unremovable
unless empty. So instead of checking if the lockfile is still used,
and removing it only if not (which is a non-atomic opration that
requires the locking on its own), we just try to remove the directory,
and that attempt succeeds only when there are no more lockers inside.

The creation operation also much simplified: instead of taking a
dir lock to make sure file is created and locked atomically, we
can just try to create everything in a loop. As soon as everything
is created, no one can remove our dir as it has the uniquely-named
tmpfile within. So uniquely-named tmpfile serves as a mandatory
lock for a dir.