shlock: try to make it fully atomic again
As shlock no longer uses the locks internally, it is now a bit
of a challenge to make it fully atomic. This is not strictly
needed, as we can rely on caller's locks, but its still a good
thing to do.

This patch changes the creation loop to never introduce the empty
lock dir. Instead the unique dir is created and populated with
the lock file, after which, this non-empty dir gets renamed to the
needed destination. If the destination already exists and is a
non-empty dir, then we move the lock file alone into it.
This avoids the possibility of the unlock code to be tricked
into removing the just-created empty dir if someone else have
stolen its own dir.