setup: set vbios ivecs not only for !vbios_post
While vbios_post should init the needed vectors in IVT, qemu's
vbios seems to be smart and recognizes the fact that it was already
inited. And doesn't set up IVT.

The proper fix might be to copy the vbios image from ROM BAR.
But ROM BAR is not always present.