dpmi: fix 0x702 impl [fixes #2100]
Use MADV_COLD instead of MADV_DONTNEED, as MADV_DONTNEED is destructive.
Also add the check to not exceed dpmi memory by that and 0x703.

Stop using MADV_FREE for 0x703, as this is unsafe unless we properly
validate the being freed region to make sure it wholly belongs to this
client.