Drop "pseudo" serial devices when fd is closed
This sets the modem status lines for a "pseudo" serial port to zero
when the associated file descriptor is closed. This is useful when
using DOSEMU2 to launch BBS doors and other programs designed to
communicate over a modem; it allows such programs to detect when the
other side of the connection has been closed.