vtmr: fix race with irr [#921]
Latch port was setting pirr without checking the result of
clearing irr first... That could lead to a race with vtmr_smi thread.

Note that the Ack port was doing the same, but it was probably
harmless there, whereas the Latch port could really lead to trouble.
Anyway, all checks are added, mutex protection enforced.