priv: dont flip uid [#2181]
We can rely on saved-euids instead, which simplifies the code a lot.
That way we need to only change euid.
setresuid() for dropping privs is avoided by the special setreuid()
trick: it always resets saved-euid if the first arg is not -1 (even
if the value is equal to current uid).