simx86: undo part of d15ace47 [#2196]
Not calling Exec_x86_fast() should be enough because the TF flag
is not set by any compiled code node: POPF and IRET are interpreted.
We may still omit calling the linker as an optimization, but that
shouldn't matter as the nodes not ended with JUMP, are not linked anyway.