pcl: dont save old stack when returning from fiber
We need to save old stack only to retrieve the stack for main
coroutine. But we never return from main coroutine. So it is enough
to save the stack only when the new coroutine is started, as there
it may be started by the main one.

See https://github.com/google/sanitizers/issues/1760