implement vlog logging subsystem [#2181]
This subsystem has the early_log buffer, so it is now possible
to log also the messages from priv.c, which are appearing before
the config and logging init.

Also this replaces the spaghetti-alike buggy code from utilities.c
with a simple (and hopefully secure) impl.
The initial log size is hard-coded to 16Kb, and the log file size
is hard-coded to 256Mb. Both can be unhardcoded, if the need appear.