group tmpfiles under single dir in tmp [#2181]
/tmp has sticky bit, so if we changed the user, we can't remove
anything from /tmp. Now only `dosemu2_<uid>_<uid>` dir is left, and
its entirely empty. It can be re-used by other dosemu2 sessions of
the same user.

Note: `dosemu2_<uid>_<uid>` is now group-writable. This is unintended,
and is a side-effect of applying ACL:
https://unix.stackexchange.com/questions/65888/setfacl-incorrectly-changes-group-permissions/498907

It basically seems like ACL creates a "fake" group with the only user
dosemu2, and for that fake group it is writable, but not for the owner's
group... It seems, after applying ACL, the only way to see the actual
group perms is via `getfacl`, and you can't even change them any more
with `chmod`. So ACL completely "shadows" the group perms, making things
really confusing.

We use ACL because dosemu2 user is not in a group of an original user.
The opposite is true: the original user should be in a dosemu2 group if
he wants to be able to remove stalled dosemu2 files in /tmp.