tcp: implement listen [closes #167]
accept() is being done on a status query.
This was yet another undocumented bit which had to be rev-enged.

Note: with slirp networking accept() doesn't work. So this driver
is now superior to our default setup where ftp2.exe couldn't get
any data (was possible to get it working with TAP setup though).