vgaemu: try different vretrace scheme [#2303]
Jazz Jackrabbit waits for vretrace and updates display_start in
vsync. So we need to synchronize this with the renderer.
Unfortunately due to high latencies, the game runs out of the
vretrace window, still producing artifacts.
Also making the vretrace window too long doesn't suit, as it makes
the game slow: it detects vretrace start by first checking for
no-retrace and then for retrace. So if the retrace window is
artificially long...