dpmi: fix off-by-1 in dpmi_is_valid_range()
PAGE_ALIGN() should only be applied to sizes.
When applying it to end_addr, we need to use end_addr+1 or addr+len.
PAGE_ALIGN(addr+len-1) is wrong, PAGE_ALIGN(end_addr) is either.

Rename "lim" to "lim_p_1" in msdos_ldt.c to denote that the calc
there was correct.