jit: fix protection of the first node [#2316]
e_mprotect() was called after e_markpage().
As such, for the first code node the mpmap entry was not yet
created, so e_markpage() just returned w/o marking anything.
We got the "ghost" node that would eventually be removed by
GC or full-page invalidation.

Fix that and add assert for mpmap entries to always exist.