simx86: cpatch writes micro-optimizations [fixes #2376]
Use querymprot() to detect slow path earlier.
On fast path try to bypass aliasmap lookup, as with querymprot()
we can hope not to fault on a write.

This noticeably speeds up zx_emul, as the slow path that it hits,
now avoids the ROM/VGA/LDT checks all together.
querymprot()+querymark() checks are rather cheap, especially for
a single-byte range.