jit: check basemode match instead of cur mode
GoodNode() used to check current mode, but this is plain wrong,
as it can be overriden by prefix, resulting in a node invalidation.
We only need to check if basemode changed, to cacth things like
PM/RM switch or 32/16bit segment switch.
Yes, segment bitness switch should invalidate the node, but the
mere prefix override - not.

Hope this fix would speed up thing quite a lot.