vgaemu: fix instremu prot in text mode [fixes #2429]
In text mode, vga.mem.map[0] points to b8000, but for instremu
we need to protect from a0000, as this is where the planes are
mapped. While the b8000 is permanently mapped to the video plane
0, a0000 can be mapped to any plane, similar to gfx mode.
This is used by edit.com to map plane 2 with text fonts, and edit
them in real time.