simx86: disable speculative prejitter again
Too problematic.
Also it is likely unsafe with cpatch, as cpatch needs to know
the exact mprot state while prejitter may concurrently change it.
Shutting down prejitter on every cpatch would be very slow and/or
you need to cut (or remove) partially prejitted nodes.

It was a great experiment and an awesome PoC, but I'd better
disable it before CI breaks again.