dpmi: honour IF when iret done on ring-0
This allows our dpmisel.S code not to care about IF skipping,
which should normally be handled by manually checking IF on stack
and conditionally calling STI. It is however difficult, as in
dpmisel.S the client bitness is not known, so we don't know the
eflags image offset on stack.

This patch makes the assumption of 2769334ff actually valid,
which was otherwise a regression.