configure: fixes for MacOS
Use cross-compile infrastucture for clang-as-as for XAS on MacOS
even on x86-64 since otherwise we end up with unusable Mach-O object
files for intermediate objects for the bios.

We can use ld.lld (via `brew install lld`) instead of i386-elf-ld but
plain ld doesn't work for linking the bios, so error out if neither
of the first two is found.