smalloc: uncommit combined areas
Uncommiting single areas doesn't actually work, as they may never
cover the full page. We need to uncommit the areas after combining.

Note that this is asymmetric with sm_commit(), because sm_commit()
always aligns the region to cover at least 1 page. That is why this
patch moves the accounting away from sm_uncommit() (but not from
sm_commit()).