Fix setjmp/longjmp from Gen_Sim PF exceptions
Unfortunately it isn't quite correct to return to Gen_Sim, as e.g. rep movsw
has a series of sim_read/writes, any of which may fail, and we need to catch
the first, not just detect TheCPU.err2 after the reads/writes are done.
Also needs to wait for the spec prejitter (if enabled) to finish as otherwise
we can't touch TheCPU.err and LONG_CS (both are needed in _Interp86).