simx86: fix rep scas/cmps cpatch code
I now actually tested the code by forcing a patch.
The trick of putting a "call (%ebx)" in front of the rep (and analyzing
the type of rep code in the patch) works ok for rep stos/movs but not
for cmps/scas since the repz/repnz after the call will still do at least one
iteration if cx>0 independently of the zero flag. So the patch needs to
insert a jmp over the rep instruction.