simx86: fix SINGLESTEP mode [fixes #2548]
It was using unexported avltr_delete()

Also in case of CODE_FLUSH(), the logging was messed up:
the instruction that was disassembled and caused CODE_FLUSH()
was already logged, but Exec() didn't yet execute it, as it
executed pending instructions. Now SINGLESTEP was printing the
registers after Exec(), but they were not matching the already
logged but not yet executed instruction, so the log got out
of order.