simx86: reorganize SynCPU structure.
The idea is to only use positive byte offsets for sim, which could
eliminate adding 128 to every offset (later). The negative offsets
are still useful for shorter calls to cpatch functions but are now
only used by the JIT.

The segment registers are now all put together in CPU encoding order,
so in previous commits any high word access had to be eliminated.
The segment descriptor caches only have a base and a base+limit.
Some other fields were never used at all, or never by offset,
sometimes as a remnant of the olden days when  a large part
was memcpy'd to and from an i386 sigcontext struct.