simx86: reorganize rep_stub.
It's now put in the array with the other stubs, and to handle things
uniformly without taking up too much space, pass an argument with
the original opcode combined with two bits for operand prefix and
repne/repe on the stack. The stack argument avoids the analysis
of the surrounding code and the extra jmp for scas/cmps.