simx86: use Gen for jmp/call far absolute in protected mode
The trick for the helper is to return the old CS, and -1 in case
an exception was set. Then the old CS can be pushed if needed
after the new one has been checked for validity and set.

Do the same thing in A_SR_SH4 so the sequence in interp.c can be
more similar for rm and pm.