simx86: do not mask not-taken branch, check limit instead
Unconditional jumps can legally be at the end of a segment, but then
the next instruction is out of bounds. A not-taken branch should
thus not be masked. However if the jump (any kind) straddles the
limit we should GPF ASAP. That can be done by returning early
and let the check at the bottom of InterpOne generate it.

Fixes #2575