simx86: remove "not_implemented" case in interp.c
cpuemu advertises a Pentium ~133-200 MHz in CPUID without MMX so all
other instructions should SIGILL (#UD) instead of leaving with a cryptic
"EMU86: error 2".

WRMSR and RDMSR should GPF instead as CPUID claims to support them but
there are no MSRs.

Fixes #1068, where NDN causes a SIGILL in vm86 using the opcode 0F 3F 07
which isn't implemented on any x86 to date and not an official UD
instruction either. It has been used in the past to detect VirtualPC,
as "VPCEXT", however.