simx86: let interrupted jumps (because of signal) follow jumps
With CKSIGN conditional jumps would return with the original
PC but it's nicer to return with the new PC, since that will
more likely be the start of another block. It also makes sure
a jump following inhibiting instructions such as pop ss, sti,
etc will complete, so signals aren't checked too early.