simx86: move POPF/IRET simulation to Sim_helper
These ones are a little more complicated in conjunction with the trap flag.
If POPF sets TF it needs to exit immediately; IRET exits via the JumpGen
mechanism. A new EXCP_SETTF pseudo exception handles this, but it works
a bit like EXCP_RETRY: it should only exit the execution loops, not
go back to the code in cpu-emu.c. The logic for this can be cleaned up
in the future with better separation.

Other flag handling is unchanged.