simx86: reorganize NewIMeta/CurrIMeta
As all instructions are compiled now we can call NewIMeta()
unconditionally at the start of InterpOne(), instead of at
multiple places when an instruction is ready. Then CurrIMeta==0
is no longer special, it's just the first instruction in the
sequence being compiled.

Some postprocessing then needed to be moved to ProduceCode().

Unlike in #2521 the tail code no longer gives a new IMeta,
to allow having only a single NewIMeta() call with overflow
backup in InterpOne(). I tested overflow again with MAXINODES=10.