simx86: use one single step for STI hold off
We can use the same mechanism as used for the trap flag to force executing
just one instruction past STI and only then exiting for PIC if needed.
We need this only for STI, not for POPSS/MOVSS. This reintroduces CeS_STI
though it doesn't directly cause CeS_RPIC to be set in HandleEmuSigals(),
it just forces one single step, and to ignore signals/pic at the end of
DoExec().

Tested with win31 and Warcraft2.