simx86: eliminate direct use of TheCPU.cs from interpreter.
We should pass it around using a variable (`ocs`) instead to be
correct with SPEC_PREJIT. Luckily it was only used in debug printfs.
I verified the asm makes sense with SPEC_PREJIT and then saw
I had accidentally enabled recursive SPEC_PREJIT, so that's gone
again, by inlining do_flush() and checking for it.