simx86: generate code on demand
We used to execute code when needed (when flushing) but now we
can do this the other way around: if no code is found, generate
some. This allows passing the just generated code node straight
to DoExec easily again instead of needing to go via the tree,
and to reduce the number of nested loops by one.

FindExecCode is the function where all this work is concentrated
now (inter_pre() was merged into it). It can only exit the loop
if TheCPU.err != 0. The only odd one there was a G->seqlen check
which seemed better as an assert.