CI: Show ACL on /dev/kvm if permission denied at open
This will help with debugging #2624, as strangely the Action should
have set up the kvm permissions in the previous step, however we get
permission denied when we try to use it.

1/ Open the /dev/kvm device file ourselves as 'kvm-ok' doesn't return
an error if the machine is kvm capable but we don't have permission
to use it.

2/ If we see permission denied, then print out the ACL on the device
file so we may glean some info as to what may be happening.